backend/requirements.minimal.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose
passlib
python-multipart
python-dotenv
email-validator
